// src/app/fullmarket/types/betslip-data.ts
export interface IBetslipData {
  backlay: string;
  sportId: string;
  matchBfId?: string;
  bfId?: string;
  marketId: string;
  matchId: string;
  runnerName: string;
  selectionId: string;
  matchName?: string;
  odds: string;
  stake: number;
  profit: number;
  loss:number;
  gameType:string
}