// src/app/shared/types/open-bets.ts
export interface IOpenBets {
  PL: number;
  betTime: string;
  betType: string;
  consolidateId: string;
  eventId: string;
  eventName: string;
  isBooked: number;
  isFancy: number;
  marketName: string;
  odds: number;
  runs: number;
  selId: number;
  selName: string;
  sportName: string;
  stake: number;
  unmatched: number;
  userName: string;
}